move practitioner controller into its own package
