+++ .gitignore
@@ -38,7 +38,6 @@ build/
 .DS_Store
 
 ### maven
-target/
 pom.xml.tag
 pom.xml.releaseBackup
 pom.xml.versionsBackup
@@ -53,3 +52,6 @@ buildNumber.properties
 
 # Test reports
 **/test-reports/
+
+# all logs folders
+**/logs
