+++ practitioner-service/src/main/java/uk/ac/ox/ndph/mts/practitioner_service/Application.java
@@ -3,21 +3,22 @@ package uk.ac.ox.ndph.mts.practitioner_service;
 /**
  * String constants
  */
-public enum Controllers {
-    
+public enum Application {
+
     // TODO: take static strings from resource file
     STARTUP("Staring practitioner service...");
-    
-    private String message;
- 
-    Controllers(String message) {
+
+    private final String message;
+
+    Application(String message) {
         this.message = message;
     }
 
     /**
-    * Returns the string associated with the enum variant.
-    * @return the string value of the enum variant
-    */
+     * Returns the string associated with the enum variant.
+     *
+     * @return the string value of the enum variant
+     */
     public String message() {
         return message;
     }
