+++ practitioner-service/src/main/java/uk/ac/ox/ndph/mts/practitioner_service/controller/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The controllers package
+ */
+package uk.ac.ox.ndph.mts.practitioner_service.controller;
